refactor: isolate todo write policy
